fix(server): full C# reference parity audit — 37 fixes across 3 sprints
Sprint 1: crystal mining every hit, movement exp, boom/protector/raz/C190
rewrite, gun charge/owner, sand gate pass-through + random diagonal,
build durability stacking, skill effect formulas, speed packet.

Sprint 2: mine exp pre-multiplier, gun damage→charge order, poli not
consumed, geopack rock pickup, C190 fx/endpoint fixes.

Sprint 3: Health skill exp + hurt FX in hurt_player_pure, aoe_damage,
C190 and gun firing system. All CLAUDE.md files updated (AGENTS→CLAUDE
refs, key files, patterns).

+++ server/net/session/outbound/player_sync.rs
@@ -14,7 +14,9 @@ pub fn send_player_speed(tx: &mpsc::UnboundedSender<Vec<u8>>, skills: &PlayerSki
     //                 100000)`
     // where `p.pause = (int)(Movement.Effect * 100)`.
     let move_effect = get_player_skill_effect(&skills.states, SkillType::Movement);
-    let pause = move_effect * 100.0;
+    // D24: C# does `int pause = (int)(Movement.Effect * 100)` — truncate to int first.
+    #[allow(clippy::cast_possible_truncation)]
+    let pause = f64::from((move_effect * 100.0) as i32);
     let xy_pause = (pause * 5.0 * 1.4 / 1000.0 * 1.7) as i32;
     let road_pause = (pause * 0.80 * 5.0 * 1.4 / 1000.0 * 1.7) as i32;
     send_u_packet(tx, "sp", &speed(xy_pause, road_pause, 100000).1);
